# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Pulse - Distributed Job Queue SaaS API",
    "version": "1.0.0",
    "description": "REST API for the Pulse distributed job queue and worker orchestration platform."
  },
  "servers": [
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Local development"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user",
        "operationId": "authRegister",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "User registered" },
          "409": { "description": "Email already exists" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login",
        "operationId": "authLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Refresh access token",
        "operationId": "authRefresh",
        "responses": {
          "200": { "description": "Token refreshed" },
          "401": { "description": "Invalid refresh token" }
        }
      }
    },
...
    "/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get current user",
        "security": [{ "bearerAuth": [] }],
        "operationId": "authMe",
        "responses": { "200": { "description": "User info" } }
      }
    },
...
    "/organizations": {
      "post": {
        "tags": ["Organizations"],
        "summary": "Create organization",
        "security": [{ "bearerAuth": [] }],
        "operationId": "createOrganization",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateOrganizationRequest" }
            }
          }
        },
        "responses": {
          "201": { "description": "Organization created" },
          "409": { "description": "Slug already exists" }
        }
      },
      "get": {
        "tags": ["Organizations"],
        "summary": "List organizations",
        "security": [{ "bearerAuth": [] }],
        "operationId": "listOrganizations",
        "responses": { "200": { "description": "List of organizations" } }
      }
    },
    "/organizations/{slug}": {
      "get": {
        "tags": ["Organizations"],
        "summary": "Get organization",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "getOrganization",
        "responses": {
          "200": { "description": "Organization details" },
          "404": { "description": "Not found" }
        }
      },
      "put": {
        "tags": ["Organizations"],
        "summary": "Update organization",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "updateOrganization",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateOrganizationRequest" }
            }
          }
        },
        "responses": {
          "200": { "description": "Organization updated" },
          "403": { "description": "Forbidden" }
        }
      },
      "delete": {
        "tags": ["Organizations"],
        "summary": "Delete organization",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "deleteOrganization",
        "responses": { "204": { "description": "Deleted" } }
      }
    },
    "/organizations/{slug}/members": {
      "get": {
        "tags": ["Organizations"],
        "summary": "List members",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "listMembers",
        "responses": { "200": { "description": "Member list" } }
      },
      "post": {
        "tags": ["Organizations"],
        "summary": "Invite member",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "inviteMember",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/InviteMemberRequest" }
            }
          }
        },
        "responses": { "201": { "description": "Member invited" } }
      }
    },
    "/organizations/{slug}/members/{memberId}": {
      "put": {
        "tags": ["Organizations"],
        "summary": "Update member role",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "memberId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "updateMemberRole",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateMemberRoleRequest" }
            }
          }
        },
        "responses": { "200": { "description": "Role updated" } }
      },
      "delete": {
        "tags": ["Organizations"],
        "summary": "Remove member",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "memberId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "removeMember",
        "responses": { "204": { "description": "Member removed" } }
      }
    },
...
    "/organizations/{slug}/projects/{projectSlug}/queues/{name}/jobs": {
      "post": {
        "tags": ["Jobs"],
        "summary": "Create job",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "projectSlug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "createJob",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateJobRequest" }
            }
          }
        },
        "responses": { "201": { "description": "Job created" } }
      },
      "get": {
        "tags": ["Jobs"],
        "summary": "List jobs",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "projectSlug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "type", "in": "query", "schema": { "type": "string" } }
        ],
        "operationId": "listJobs",
        "responses": { "200": { "description": "Paginated jobs" } }
      }
    },
    "/organizations/{slug}/projects/{projectSlug}/queues/{name}/jobs/{jobId}": {
      "get": {
        "tags": ["Jobs"],
        "summary": "Get job",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "projectSlug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "getJob",
        "responses": { "200": { "description": "Job details with logs" } }
      },
      "delete": {
        "tags": ["Jobs"],
        "summary": "Delete job",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "projectSlug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "deleteJob",
        "responses": { "204": { "description": "Deleted" } }
      }
    },
    "/organizations/{slug}/projects/{projectSlug}/queues/{name}/jobs/{jobId}/retry": {
      "post": {
        "tags": ["Jobs"],
        "summary": "Retry failed job",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "projectSlug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "retryJob",
        "responses": { "200": { "description": "Job queued for retry" } }
      }
    },
    "/organizations/{slug}/projects/{projectSlug}/queues/{name}/jobs/{jobId}/cancel": {
      "post": {
        "tags": ["Jobs"],
        "summary": "Cancel job",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "projectSlug", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "name", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "operationId": "cancelJob",
        "responses": { "200": { "description": "Job cancelled" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "password", "firstName", "lastName"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "minLength": 8 },
          "firstName": { "type": "string" },
          "lastName": { "type": "string" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
      "CreateOrganizationRequest": {
        "type": "object",
        "required": ["name", "slug"],
        "properties": {
          "name": { "type": "string" },
          "slug": { "type": "string", "pattern": "^[a-z0-9-]+$" }
        }
      },
      "UpdateOrganizationRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string" }
        }
      },
      "InviteMemberRequest": {
        "type": "object",
        "required": ["email", "role"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "role": { "type": "string", "enum": ["OWNER", "ADMIN", "DEVELOPER", "VIEWER"] }
        }
      },
      "UpdateMemberRoleRequest": {
        "type": "object",
        "required": ["role"],
        "properties": {
          "role": { "type": "string", "enum": ["OWNER", "ADMIN", "DEVELOPER", "VIEWER"] }
        }
      },
...
      "CreateJobRequest": {
        "type": "object",
        "required": ["payload"],
        "properties": {
          "type": { "type": "string", "enum": ["IMMEDIATE", "DELAYED", "SCHEDULED", "RECURRING", "BATCH"] },
          "payload": { "type": "object" },
          "priority": { "type": "integer" },
          "maxRetries": { "type": "integer" },
          "cronExpression": { "type": "string" },
          "runAt": { "type": "string", "format": "date-time" },
          "idempotencyKey": { "type": "string" },
          "batchId": { "type": "string" },
          "parentJobId": { "type": "string" }
        }
      },
...
    }
  }
}